Remove InputLoopService from pubsub/native Core_9 sample and integrate logic into Program.cs
