+++ samples/pubsub/native/Core_9/Publisher/Program.cs
@@ -3,17 +3,48 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using NServiceBus;
-using Publisher;
 
 Console.Title = "Publisher";
 var builder = Host.CreateApplicationBuilder(args);
-builder.Services.AddHostedService<InputLoopService>();
 
 var endpointConfiguration = new EndpointConfiguration("Samples.PubSub.Publisher");
 endpointConfiguration.UseSerialization<SystemJsonSerializer>();
 endpointConfiguration.UseTransport(new LearningTransport());
-Console.WriteLine("Press any key, application loading");
-Console.ReadKey();
-Console.WriteLine("Starting...");
+
 builder.UseNServiceBus(endpointConfiguration);
-await builder.Build().RunAsync();
+
+var host = builder.Build();
+
+await host.StartAsync();
+
+var messageSession = host.Services.GetRequiredService<IMessageSession>();
+
+Console.WriteLine("Press '1' to publish the OrderReceived event");
+Console.WriteLine("Press any other key to exit");
+
+#region PublishLoop
+
+while (true)
+{
+    var key = Console.ReadKey();
+    Console.WriteLine();
+
+    var orderReceivedId = Guid.NewGuid();
+    if (key.Key == ConsoleKey.D1)
+    {
+        var orderReceived = new OrderReceived
+        {
+            OrderId = orderReceivedId
+        };
+        await messageSession.Publish(orderReceived);
+        Console.WriteLine($"Published OrderReceived Event with Id {orderReceivedId}.");
+    }
+    else
+    {
+        break;
+    }
+}
+
+#endregion
+
+await host.StopAsync();
